restructure existing tests for scopy
