+++ test/scopy/windows/common.py
